Fix itself and minor refactoring
