+++ packages/victoria-lens/src/App.tsx
@@ -4,40 +4,8 @@ import {MetricResult, QueryRangeResponse} from "./api/types";
 import GraphView from "./components/GraphView";
 import {AppBar, Box, Grid, TextField, Toolbar, Typography} from "@material-ui/core";
 import {ToggleButton, ToggleButtonGroup} from "@material-ui/lab";
-import {TimeParams} from "./types";
 import {getQueryRange} from "./api/query-range";
-
-export enum TimePreset {
-  "lastHour",
-  "last15Min",
-  "last2Min",
-  "last24Hours"
-};
-
-const getTimeperiodForPreset = (p: TimePreset): TimeParams => {
-  const n = (new Date()).valueOf() / 1000;
-  let delta = 0;
-  switch (p) {
-    case TimePreset.last2Min:
-      delta = 60 * 2;
-      break;
-    case TimePreset.last15Min:
-      delta = 60 * 15;
-      break;
-    case TimePreset.lastHour:
-      delta = 60 * 60;
-      break;
-    case TimePreset.last24Hours:
-      delta = 60 * 60 * 24;
-  }
-
-  const MAX_ITEMS_PER_CHART = 300; // TODO: make dependent from screen size
-  return {
-    start: n - delta,
-    end: n,
-    step: delta / MAX_ITEMS_PER_CHART
-  }
-}
+import {getTimeperiodForPreset, TimePreset} from "./utils/time";
 
 function App() {
 
@@ -63,16 +31,6 @@ function App() {
       }})()
   }, [timePreset]);
 
-  const onSubmitHandler = async () => {
-    let response = await fetch(`${server}/api/v1/query_range?query=${query}&start=${period.start}&end=${period.end}&step=15s`);
-    if (response.ok) {
-      const resp: QueryRangeResponse = await response.json();
-      setGraphData(resp.data.result);
-    } else {
-      alert((await response.json())?.error);
-    }
-  };
-
   return (
       <>
         <AppBar position="static">
@@ -91,7 +49,11 @@ function App() {
               <ToggleButtonGroup
                   value={timePreset}
                   exclusive
-                  onChange={(e, val) => setTimePreset(val)}>
+                  onChange={
+                    (e, val) =>
+                        // Toggle Button Group returns null in case of click on selected element, avoiding it
+                        setTimePreset(prev => val ?? prev)
+                  }>
                 <ToggleButton value={TimePreset.last2Min} aria-label="last 2 min">
                   2 min
                 </ToggleButton>
